Refactored Dash app into a modular structure
